\name{createArea}
\alias{createArea}
\title{A Function To Create The Reservoirs System Area}
\description{it creates an area to operate reservoirs inside it}
\usage{createArea(name = "unknown", location = "unknown", start = c(), end = c())}
\arguments{
  \item{name}{a string representing the name of the operating area}
  \item{location}{a string representing the location of the operating area}
  \item{start}{a vector whose first number presents the number of start month of simulation time and second number shows the start year of of simulation time}
  \item{end}{a vector whose first number presents the number of end month of simulation time and second number shows the end year of of simulation time}
}
\value{an object of class of createArea}
\author{Rezgar Arabzadeh, Parisa Aberi}
\keyword{ models }