\name{object}
\alias{object}
\docType{data}
\title{An Object From Class Of creatArea}
\description{an object from the class of createArea which includes the specifications of a cascde reservoir system in Kurdistan, ZerineRud basin. "object" includes two inflow time series in the Bukan and Sonata dam sites, there demands sites time series and geometrical specifications of mentioned dams}
\usage{data("object")}
\references{
Water Resources Management Company, Ministry of Energy, Iran, 2015}
\keyword{datasets}
