\name{plot.SOP}
\alias{plot.SOP}
\title{A Plot Function for Object Inherited From Class Of SOP }
\description{a plot function for an object inherited from class SOP}
\usage{
\method{plot}{SOP}(x , ...)
}
\arguments{
  \item{x}{an object inherited from class SOP}
  \item{...}{other objects that can passed to plot function}
}
\author{Rezgar Arabzadeh, Parisa Aberi, Kaveh Panaghi}
\keyword{ aplot }