\name{RSSOP-package}
\alias{RSSOP-package}
\docType{package}
\title{Tools for Operation of Reservoirs Systems using Standard Operation Policy}
\description{RSSOP}
\details{
\tabular{ll}{
Package: \tab RSSOP\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-12-29\cr
License: \tab GPL-2\cr
}
The package provides functions to establish the reservoirs systems elements and build up an operation standard operation policy (SOP) releases.
The package include S3 classes for reservoir(s) system operation. 
Methods such as plot and yield is available for standard objects inherited from class SOP for results illustration and visualization.}
\author{
Rezgar Arabzadeh <rezgararabzadeh@ut.ac.ir>,
Parisa Aberi <ParisaAberi@gmail.com>,
Kaveh Panaghi <k.panaghi@gmail.com>,
Shahab Araghinejad <araghinejad@ut.ac.ir,
Majid Montaseri <montaseri@hotmail.com>

Maintainer: Rezgar Arabzadehh <rezgararabzadeh@ut.ac.ir>
}
\references{
Loucks, Daniel P., et al. Water resources systems planning and management: an introduction to methods, models and applications. Paris: UNESCO, 2005.
Araghinejad, S. (2013). Data-Driven Modeling: Using MATLAB, in Water Resources and Environmental Engineering (Vol. 67). Springer Science & Business Media.
}
\keyword{ package }
