\name{SOP.default}
\alias{SOP.default}
\title{Default Function Closs Of SOP}
\description{SOP default function of class SOP}
\usage{
\method{SOP}{default}(object)
}
\arguments{
  \item{object}{an object inherited from class createArea}
}
\author{Rezgar Arabzadeh, Parisa Aberi}
\keyword{ models }