\name{Yeild}
\alias{Yeild}
\title{Reservoir Performance Indices}
\description{a function for evaluation and calculation of reservoir performance indices based on Hashimoto et al. (1982)}
\usage{
Yeild(object, s.const = 0.95)
}
\arguments{
  \item{object}{an object inherited from class SOP}
  \item{s.const}{satisfactory constant of supplying}
}
\value{a matrix presenting Reliability, resiliency, and vulnerability criterion for water resource system performance evaluation}
\references{
Hashimoto, T., Stedinger, J. R., & Loucks, D. P. (1982). Reliability, resiliency, and vulnerability criteria for water resource system performance evaluation. Water resources research, 18(1), 14-20.
}
\author{Rezgar Arabzdadeh, Parisa Aberi}
\keyword{ models }